languages/python/beginner-projects/14-ocr-pipeline-for-scanned-documents/requirements.txt
pytesseract>=0.3.13
Pillow>=10.0.0
